requirements.txt
python-telegram-bot==22.5
langchain
langgraph
langchain-openai
httpx
pypdf
python-dotenv
apscheduler
wikipedia
